+++ default.nix
@@ -19,7 +19,7 @@ ghc."8.10" = nixos-22-05.haskell.packages.ghc8107.override (old: {
         (new: old: {
             ascii-case = new.callPackage ./nix/ascii-case-1.0.1.0.nix {};
             ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
-            ascii-superset = new.callPackage ./nix/ascii-superset-1.1.0.0.nix {};
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.2.4.0.nix {};
         })
     ];
 });
@@ -30,7 +30,7 @@ ghc."9.0" = nixos-22-11.haskell.packages.ghc90.override (old: {
         (new: old: {
             ascii-case = new.callPackage ./nix/ascii-case-1.0.1.0.nix {};
             ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
-            ascii-superset = new.callPackage ./nix/ascii-superset-1.1.0.0.nix {};
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.2.4.0.nix {};
         })
     ];
 });
@@ -41,7 +41,7 @@ ghc."9.2" = nixos-22-11.haskell.packages.ghc92.override (old: {
         (new: old: {
             ascii-case = new.callPackage ./nix/ascii-case-1.0.1.0.nix {};
             ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
-            ascii-superset = new.callPackage ./nix/ascii-superset-1.1.0.0.nix {};
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.2.4.0.nix {};
         })
     ];
 });
@@ -52,7 +52,7 @@ ghc."9.4" = nixos-22-11.haskell.packages.ghc94.override (old: {
         (new: old: {
             ascii-case = new.callPackage ./nix/ascii-case-1.0.1.0.nix {};
             ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
-            ascii-superset = new.callPackage ./nix/ascii-superset-1.2.0.0.nix {};
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.2.4.0.nix {};
         })
     ];
 });